backend/requirements.txt
# FastAPI и основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.11
redis==5.0.1

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-decouple==3.8

# HTTP клиенты
httpx==0.25.2
aiohttp==3.9.1

# AI и ML
openai==1.3.7
celery==5.3.4

# Логирование и мониторинг
loguru==0.7.2
prometheus-client==0.19.0

# Валидация данных
pydantic==2.9.2
pydantic-settings==2.5.2

# Утилиты
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Разработка
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
